Use transformers library for SAM/GroundingDINO
