+++ plantclef/masking/transform.py
@@ -2,17 +2,21 @@ import io
 import numpy as np
 import torch
 from PIL import Image
-from segment_anything import SamPredictor, sam_model_registry
-from groundingdino.util.inference import Model
+from transformers import (
+    SamModel,
+    SamProcessor,
+    AutoProcessor,
+    AutoModelForZeroShotObjectDetection,
+)
 from typing import List
-from plantclef.model_setup import (
-    setup_segment_anything_checkpoint_path,
-    setup_groundingdino_checkpoint_path,
-    setup_groundingdino_config_path,
+
+from .params import (
+    HasCheckpointPathSAM,
+    HasCheckpointPathGroundingDINO,
+    HasBatchSize,
 )
 
 from pyspark.ml import Transformer
-from pyspark.ml.param import Param, Params, TypeConverters
 from pyspark.ml.param.shared import HasInputCol, HasOutputCol
 from pyspark.ml.util import DefaultParamsReadable, DefaultParamsWritable
 from pyspark.sql import DataFrame
@@ -20,124 +24,12 @@ from pyspark.sql import functions as F
 from pyspark.sql.types import StructType, StructField, BinaryType
 
 
-class HasCheckpointPathSAM(Param):
-    """
-    Mixin for param checkpoint_path: str
-    """
-
-    checkpointPathSAM = Param(
-        Params._dummy(),
-        "checkpointPathSAM",
-        "The path to the segment-anything checkpoint weights",
-        typeConverter=TypeConverters.toString,
-    )
-
-    def __init__(self):
-        super().__init__(
-            default=setup_segment_anything_checkpoint_path(),
-            doc="The path to the segment-anything checkpoint weights",
-        )
-
-    def getCheckpointPathSAM(self) -> str:
-        return self.getOrDefault(self.checkpointPathSAM)
-
-
-class HasEncoderVersion(Param):
-    """
-    Mixin for param encoder_version: str
-    """
-
-    encoderVersion = Param(
-        Params._dummy(),
-        "encoderVersion",
-        "The name of the SAM encoder version to use",
-        typeConverter=TypeConverters.toString,
-    )
-
-    def __init__(self):
-        super().__init__(
-            default="vit_h",
-            doc="The name of the SAM encoder version to use",
-        )
-
-    def getEncoderVersion(self) -> str:
-        return self.getOrDefault(self.encoderVersion)
-
-
-class HasCheckpointPathGroundingDINO(Param):
-    """
-    Mixin for param checkpoint_path_groundingdino: str
-    """
-
-    checkpointPathGroundingDINO = Param(
-        Params._dummy(),
-        "checkpointPathGroundingDINO",
-        "The path to the GroundingDINO checkpoint weights",
-        typeConverter=TypeConverters.toString,
-    )
-
-    def __init__(self):
-        super().__init__(
-            default=setup_groundingdino_checkpoint_path(),
-            doc="The path to the GroundingDINO checkpoint weights",
-        )
-
-    def getCheckpointPathGroundingDINO(self) -> str:
-        return self.getOrDefault(self.checkpointPathGroundingDINO)
-
-
-class HasConfigPathGroundingDINO(Param):
-    """
-    Mixin for param config_path_groundingdino: str
-    """
-
-    configPathGroundingDINO = Param(
-        Params._dummy(),
-        "configPathGroundingDINO",
-        "The path to the GroundingDINO config path",
-        typeConverter=TypeConverters.toString,
-    )
-
-    def __init__(self):
-        super().__init__(
-            default=setup_groundingdino_config_path(),
-            doc="The path to the GroundingDINO config path",
-        )
-
-    def getConfigPathGroundingDINO(self) -> str:
-        return self.getOrDefault(self.configPathGroundingDINO)
-
-
-class HasBatchSize(Param):
-    """
-    Mixin for param batch_size: int
-    """
-
-    batchSize = Param(
-        Params._dummy(),
-        "batchSize",
-        "The batch size to use for embedding extraction",
-        typeConverter=TypeConverters.toInt,
-    )
-
-    def __init__(self):
-        super().__init__(
-            default=32,
-            doc="The batch size to use for embedding extraction",
-        )
-
-    def getBatchSize(self) -> int:
-        return self.getOrDefault(self.batchSize)
-
-
 class WrappedMasking(
     Transformer,
     HasInputCol,
     HasOutputCol,
     HasCheckpointPathSAM,
-    HasEncoderVersion,
     HasCheckpointPathGroundingDINO,
-    HasConfigPathGroundingDINO,
     HasBatchSize,
     DefaultParamsReadable,
     DefaultParamsWritable,
@@ -150,32 +42,29 @@ class WrappedMasking(
         self,
         input_col: str = "data",
         output_col: str = "masks",
-        checkpoint_path_sam: str = setup_segment_anything_checkpoint_path(),
-        checkpoint_path_groundingdino: str = setup_groundingdino_checkpoint_path(),
-        config_path_groundingdino: str = setup_groundingdino_config_path(),
-        encoder_version: str = "vit_h",
+        checkpoint_path_sam: str = "facebook/sam-vit-huge",
+        checkpoint_path_groundingdino: str = "IDEA-Research/grounding-dino-base",
         batch_size: int = 32,
     ):
         super().__init__()
         self._setDefault(
             inputCol=input_col,
             outputCol=output_col,
+            # NOTE: these are more ids than they are checkpoint paths
             checkpointPathSAM=checkpoint_path_sam,
-            encoderVersion=encoder_version,
             checkpointPathGroundingDINO=checkpoint_path_groundingdino,
-            configPathGroundingDINO=config_path_groundingdino,
             batchSize=batch_size,
         )
         self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        # Segment Anything Model
-        sam = sam_model_registry[self.getEncoderVersion()](
-            checkpoint=self.getCheckpointPathSAM()
-        ).to(device=self.device)
-        self.sam_predictor = SamPredictor(sam)
-        # GroundingDINO Model
-        self.groundingdino_model = Model(
-            model_config_path=self.getConfigPathGroundingDINO(),
-            model_checkpoint_path=self.getCheckpointPathGroundingDINO(),
+        # https://huggingface.co/docs/transformers/main/en/model_doc/sam
+        self.sam_model = SamModel.from_pretrained(checkpoint_path_sam).to(self.device)
+        self.sam_processor = SamProcessor.from_pretrained(checkpoint_path_sam)
+        # https://huggingface.co/docs/transformers/main/en/model_doc/grounding-dino#transformers.GroundingDinoForObjectDetection
+        self.groundingdino_model = AutoModelForZeroShotObjectDetection(
+            checkpoint_path_groundingdino
+        ).to(self.device)
+        self.groundingdino_processor = AutoProcessor.from_pretrained(
+            checkpoint_path_groundingdino
         )
         # params for groundingdino
         self.CLASSES = [
@@ -209,19 +98,23 @@ class WrappedMasking(
 
     def segment(
         self,
-        sam_predictor: SamPredictor,
         image: np.ndarray,
         xyxy: np.ndarray,
     ) -> np.ndarray:
-        sam_predictor.set_image(image)
-        result_masks = []
-        for box in xyxy:
-            masks, scores, logits = sam_predictor.predict(
-                box=box, multimask_output=True
-            )
-            index = np.argmax(scores)
-            result_masks.append(masks[index])
-        return np.array(result_masks)
+        inputs = self.sam_processor(image, input_points=xyxy, return_tensors="pt").to(
+            self.device
+        )
+        with torch.no_grad():
+            outputs = self.sam_model(**inputs)
+
+        masks = self.sam_processor.image_processor.post_process_masks(
+            outputs.pred_masks.cpu(),
+            inputs["original_sizes"].cpu(),
+            inputs["reshaped_input_sizes"].cpu(),
+        )
+        # scores = outputs.iou_scores
+        # convert to numpy
+        return masks.numpy()
 
     def enhance_class_name(self, class_names: List[str]) -> List[str]:
         return [f"all {class_name}" for class_name in class_names]
@@ -250,31 +143,41 @@ class WrappedMasking(
 
         def predict(input_image: np.ndarray) -> np.ndarray:
             # convert binary to RGB
-            pil_image = Image.open(io.BytesIO(input_image)).convert("RGB")
-            image_np = np.array(pil_image)
+            image = Image.open(io.BytesIO(input_image)).convert("RGB")
 
             # predict with groundingdino
-            detections = self.groundingdino_model.predict_with_classes(
-                image=image_np,
-                classes=self.enhance_class_name(class_names=self.CLASSES),
-                box_threshold=self.BOX_THRESHOLD,
-                text_threshold=self.TEXT_THRESHOLD,
+            inputs = self.groundingdino_processor(
+                images=image,
+                text=self.enhance_class_name(class_names=self.CLASSES),
+                return_tensors="pt",
+            ).to(self.device)
+
+            with torch.no_grad():
+                outputs = self.groundingdino_model(**inputs)
+
+            detections = (
+                self.groundingdino_processor.post_process_grounded_object_detection(
+                    outputs,
+                    threshold=self.BOX_THRESHOLD,
+                    text_threshold=self.TEXT_THRESHOLD,
+                    target_sizes=[(image.height, image.width)],
+                )
             )
 
-            # generate masks
-            detections.mask = self.segment(
-                sam_predictor=self.sam_predictor,
-                image=image_np,
-                xyxy=detections.xyxy,
-            )
+            masks = self.segment(image=image, xyxy=detections["boxes"])
 
-            merged_masks = []
-            class_mask_results = {"leaf": None, "flower": None, "plant": None}
+            empty_image = self.empty_png(image.shape[:2])
+            class_mask_results = {
+                "leaf": empty_image,
+                "flower": empty_image,
+                "plant": empty_image,
+            }
             grouped = {}
-            for mask, class_id in zip(detections.mask, detections.class_id):
+            for mask, class_id in zip(masks, np.argmax(detections["scores"])):
                 grouped.setdefault(class_id, []).append(mask)
 
             # get the masks for the classes we are interested in
+            merged_masks = []
             for class_id, masks in grouped.items():
                 if class_id in self.INCLUDE_CLASS_IDS and len(masks) > 0:
                     merged_mask = np.any(np.stack(masks, axis=0), axis=0)
@@ -289,32 +192,11 @@ class WrappedMasking(
             )
 
             # make sure masks are in binary
-            combined = (
-                final_mask_bytes
-                if final_mask_bytes
-                else self.empty_png(image_np.shape[:2])
-            )
-            leaf = (
-                class_mask_results["leaf"]
-                if class_mask_results["leaf"]
-                else self.empty_png(image_np.shape[:2])
-            )
-            flower = (
-                class_mask_results["flower"]
-                if class_mask_results["flower"]
-                else self.empty_png(image_np.shape[:2])
-            )
-            plant = (
-                class_mask_results["plant"]
-                if class_mask_results["plant"]
-                else self.empty_png(image_np.shape[:2])
-            )
+            combined = final_mask_bytes or empty_image
 
             return {
                 "combined_mask": combined,
-                "leaf_mask": leaf,
-                "flower_mask": flower,
-                "plant_mask": plant,
+                **{f"{k}_mask": v for k, v in class_mask_results.items()},
             }
 
         return predict
